working gradle

+++ extensions/grpc/deployment/src/main/java/io/quarkus/grpc/deployment/PreBuildProcessor.java
@@ -4,18 +4,21 @@ import static java.util.Arrays.asList;
 
 import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Consumer;
+import java.util.jar.Attributes;
+import java.util.jar.JarInputStream;
+import java.util.jar.Manifest;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 import io.quarkus.bootstrap.PreBuildContext;
 import io.quarkus.bootstrap.model.AppArtifact;
-import io.quarkus.bootstrap.model.AppDependency;
-import io.quarkus.bootstrap.model.AppModel;
 import io.quarkus.bootstrap.resolver.AppModelResolver;
 import io.quarkus.bootstrap.resolver.AppModelResolverException;
 import io.quarkus.deployment.annotations.PreBuildStep;
@@ -25,7 +28,7 @@ import io.quarkus.utilities.OS;
 // mstodo test windows
 // mstodo customize via properties
 // mstodo customizable proto path
-public class PreBuildProcessor {
+public class PreBuildProcessor { // mstodo GrpcPreProcessor
     private static final String CLASS_PATH_DELIMITER = "" + File.pathSeparatorChar;
 
     private static final String quarkusProtocPluginMain = "io.quarkus.grpc.protoc.plugin.MutinyGrpcGenerator";
@@ -78,9 +81,13 @@ public class PreBuildProcessor {
                         "--java_out=" + protoOut));
                 command.addAll(protoFiles);
 
-                Process process = new ProcessBuilder()
-                        .command(command)
-                        .inheritIO().start();
+                // mstodo rollback output redirection?
+                ProcessBuilder builder = new ProcessBuilder()
+                        .command(command);
+                builder
+                        .redirectErrorStream(true)
+                        .redirectOutput(File.createTempFile("from-graddily-doo", ".txt"));
+                Process process = builder.start();
                 int resultCode = process.waitFor();
                 if (resultCode != 0) {
                     throw new RuntimeException("Failed to generate Java classes from proto file: " + protoFiles);
@@ -112,17 +119,20 @@ public class PreBuildProcessor {
 
     private Path prepareExecutable(Path buildDir, AppModelResolver resolver, AppArtifact a, String exeName) throws IOException {
         Path exeArtifact;
-        try {
-            exeArtifact = resolver.resolve(a);
-        } catch (AppModelResolverException e) {
-            throw new RuntimeException("Unable to resolve the executable: " + a +
-                    ". It may mean you are using an unsupported architecture. Consider switching" +
-                    " to maven-protoc-plugin for java classes generation", e);
-        }
         Path exe = buildDir.resolve(exeName);
-        Files.copy(exeArtifact, exe);
-        if (!exe.toFile().setExecutable(true)) {
-            throw new RuntimeException("Failed to make the file executable: " + exe);
+        if (!Files.exists(exe)) {
+            // only download the artifact if we don't have it yet:
+            try {
+                exeArtifact = resolver.resolve(a);
+            } catch (AppModelResolverException e) {
+                throw new RuntimeException("Unable to resolve the executable: " + a +
+                        ". It may mean you are using an unsupported architecture. Consider switching" +
+                        " to maven-protoc-plugin for java classes generation", e);
+            }
+            Files.copy(exeArtifact, exe);
+            if (!exe.toFile().setExecutable(true)) {
+                throw new RuntimeException("Failed to make the file executable: " + exe);
+            }
         }
         return exe;
     }
@@ -151,12 +161,19 @@ public class PreBuildProcessor {
     private static Path prepareQuarkusGrpcExecutable(AppModelResolver resolver,
             Path buildDir)
             throws IOException, AppModelResolverException {
+        List<Path> qGrpcPluginClasspath = new ArrayList<>();
+
         String quarkusGrpcVersion = PreBuildProcessor.class.getPackage().getImplementationVersion();
         AppArtifact quarkusGrpcPlugin = new AppArtifact("io.quarkus", "quarkus-grpc-protoc-plugin", quarkusGrpcVersion);
-        AppModel qGrpcPluginModel = resolver.resolveModel(quarkusGrpcPlugin);
-        String classpath = qGrpcPluginModel.getUserDependencies().stream()
-                .map(AppDependency::getArtifact)
-                .map(a -> getArtifact(resolver, a))
+        Path pluginPath = resolver.resolve(quarkusGrpcPlugin);
+        qGrpcPluginClasspath.add(pluginPath);
+        List<AppArtifact> classpathArtifacts = readClasspath(pluginPath);
+        for (AppArtifact classpathArtifact : classpathArtifacts) {
+            Path dependencypath = resolver.resolve(classpathArtifact);
+            qGrpcPluginClasspath.add(dependencypath);
+        }
+
+        String classpath = qGrpcPluginClasspath.stream()
                 .map(Path::toAbsolutePath)
                 .map(Path::toString)
                 .collect(Collectors.joining(CLASS_PATH_DELIMITER));
@@ -166,28 +183,60 @@ public class PreBuildProcessor {
         classpath += getArtifact(resolver, quarkusGrpcPlugin).toAbsolutePath().toString();
 
         Path script;
-        BufferedWriter writer;
-        if (OS.determineOS() != OS.WINDOWS) {
-            script = Files.createTempFile(buildDir, "quarkus-grpc", ".sh");
+        BufferedWriter writer = null;
+        try {
+            if (OS.determineOS() != OS.WINDOWS) {
+                script = Files.createTempFile(buildDir, "quarkus-grpc", ".sh");
 
-            writer = Files.newBufferedWriter(script);
+                writer = Files.newBufferedWriter(script);
 
-            writer.write("#!/bin/sh\n");
+                writer.write("#!/bin/sh\n");
+                writer.newLine();
+            } else {
+                script = Files.createTempFile(buildDir, "quarkus-grpc", ".cmd");
+                writer = Files.newBufferedWriter(script);
+            }
+            writer.write(JavaBinFinder.findBin() + " -cp " +
+                    classpath + " " + quarkusProtocPluginMain);
             writer.newLine();
-        } else {
-            script = Files.createTempFile(buildDir, "quarkus-grpc", ".cmd");
-            writer = Files.newBufferedWriter(script);
+        } finally {
+            if (writer != null) {
+                writer.close();
+            }
         }
-        writer.write(JavaBinFinder.findBin() + " -cp " +
-                classpath + " " + quarkusProtocPluginMain);
-        writer.newLine();
-        writer.close();
 
         if (!script.toFile().setExecutable(true)) {
             System.out.println("failed to set file: " + script + " executable. Protoc invocation may fail");
             // mstodo log instead
         }
         return script;
+    }
+
+    private static List<AppArtifact> readClasspath(Path pluginPath) {
+        List<AppArtifact> result = new ArrayList<>();
+        try (FileInputStream stream = new FileInputStream(pluginPath.toFile());
+                JarInputStream jarStream = new JarInputStream(stream)) {
+            Manifest manifest = jarStream.getManifest();
+            String classPath = (String) manifest.getMainAttributes().get(new Attributes.Name("Class-Path"));
+            String[] artifacts = classPath.split(Pattern.quote(" ")); // mstodo check multiple artifacts on the classpath
+            for (String artifact : artifacts) {
+                String[] gav_cp = artifact.split(":");
+                if (gav_cp.length < 4 || gav_cp.length > 5) {
+                    throw new RuntimeException(
+                            "Invalid artifact specifier:  " + artifact + " in the classpath of the qarkus grpc plugin");
+                }
+                String groupId = gav_cp[0];
+                String artifactId = gav_cp[1];
+                String version = gav_cp[2];
+                String classifier = gav_cp.length == 4 ? null : gav_cp[3];
+                String packaging = gav_cp.length == 4 ? gav_cp[3] : gav_cp[4];
+
+                result.add(new AppArtifact(groupId, artifactId, classifier, packaging, version));
+            }
+        } catch (IOException e) {
+            throw new RuntimeException("Failure constructing classpath"); // mstodo
+        }
+        return result;
     }
 
     private static Path getArtifact(AppModelResolver resolver, AppArtifact a) {
